+++ lib/generators/rails/USAGE
