+++ lib/generators/rails/serializer_generator.rb
