+++ lib/generators/rails/templates/serializer.rb.erb
